+++ example/lib/screens/order_page.dart
@@ -106,7 +106,7 @@
 //                                 ),
 //                                 const SizedBox(width: 15),
 //                                 Text(
-                                   // ignore: lines_longer_than_80_chars
+// ignore: lines_longer_than_80_chars
 //                              'Price: ${(order.items[i].count * order.items[i].item!.price).toStringAsFixed(2)} €',
 //                                   style: TextStyle(fontSize: 15),
 //                                 ),