+++ lib/src/controllers/middleware_controller.dart
@@ -5,8 +5,9 @@ class MiddlewareController {
   MiddlewareController(this.route);
 
   Future<String?> runRedirect() async {
+    final path = route.getLastActivePath();
     for (var middle in route.route.middleware!) {
-      final result = await middle.redirectGuard();
+      final result = await middle.redirectGuard(path);
       if (result != null) {
         return result;
       }
@@ -39,5 +40,17 @@ class MiddlewareController {
     for (var middle in route.route.middleware!) {
       middle.onMatch();
     }
+  }
+
+  bool runCanPop() {
+    if (!route.hasMiddlewares) {
+      return true;
+    }
+    for (var middle in route.route.middleware!) {
+      if (!middle.canPop()) {
+        return false;
+      }
+    }
+    return true;
   }
 }