+++ lib/src/controllers/qrouter_controller.dart
@@ -5,6 +5,7 @@ import 'package:flutter/widgets.dart';
 import '../../qlevar_router.dart';
 import '../pages/qpage_internal.dart';
 import '../qr.dart';
+import '../routes/qroute_builder.dart';
 import '../routes/qroute_children.dart';
 import '../routes/qroute_internal.dart';
 import '../types/qhistory.dart';
@@ -16,6 +17,8 @@ import 'pages_controller.dart';
 abstract class QNavigator extends ChangeNotifier {
   /// Get if the cureent [QNavigator] can pop or not
   bool get canPop;
+
+  bool get isDeclarative;
 
   /// Get the current route for this navigator
   QRoute get currentRoute;
@@ -52,16 +55,26 @@ class QRouterController extends QNavigator {
 
   final QRouteChildren routes;
 
+  final PagesBuilder? declarativePagesBuilder;
+
   final _pagesController = PagesController();
 
-  QRouterController(this.key, this.routes,
-      {String? initPath, QRouteInternal? initRoute}) {
+  QRouterController(
+    this.key,
+    this.routes, {
+    String? initPath,
+    QRouteInternal? initRoute,
+    this.declarativePagesBuilder,
+  }) {
     if (initRoute != null) {
       addRoute(initRoute);
     } else {
       push(initPath!);
     }
   }
+
+  @override
+  bool get isDeclarative => declarativePagesBuilder != null;
 
   @override
   QRoute get currentRoute => _pagesController.routes.last.route;
@@ -90,9 +103,11 @@ class QRouterController extends QNavigator {
     if (!canPop) {
       return false;
     }
-    _pagesController.removeLast();
-    update(withParams: true);
-    return true;
+    final isPoped = _pagesController.removeLast();
+    if (isPoped) {
+      update(withParams: true);
+    }
+    return isPoped;
   }
 
   @override
@@ -237,4 +252,6 @@ class QRouterController extends QNavigator {
       QR.history.removeLast();
     }
   }
+
+  void updateDeclarative({QRouteInternal? match}) {}
 }